refactoring bar chart ui
